backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
cloudinary==1.36.0 